Refactoring transaction details + bottom sheet
