+++ locales/base/translation.json
@@ -515,6 +515,7 @@
   "to": "To",
   "for": "For",
   "reclaim": "Reclaim",
+  "claimed": "Claimed",
   "celoGold": "CELO",
   "celoEuros": "Celo Euros",
   "balanceAvailable": "<0></0> available",
