+++ src/navigator/Navigator.tsx
@@ -110,6 +110,7 @@ import SwapScreen from 'src/swap/SwapScreen'
 import AssetsScreen from 'src/tokens/Assets'
 import TokenDetailsScreen from 'src/tokens/TokenDetails'
 import TokenImportScreen from 'src/tokens/TokenImport'
+import JumpstartTransactionDetailsScreen from 'src/transactions/feed/JumpstartTransactionDetailsScreen'
 import TransactionDetailsScreen from 'src/transactions/feed/TransactionDetailsScreen'
 import JumpstartReclaimBottomSheet from 'src/transactions/feed/detailContent/JumpstartReclaimBottomSheet'
 import Logger from 'src/utils/Logger'
@@ -501,6 +502,11 @@ const generalScreens = (Navigator: typeof Stack) => (
       name={Screens.TransactionDetailsScreen}
       component={TransactionDetailsScreen}
       options={headerWithBackButton}
+    />
+    <Navigator.Screen
+      name={Screens.JumpstartTransactionDetailsScreen}
+      component={JumpstartTransactionDetailsScreen}
+      options={noHeader}
     />
     <Navigator.Screen
       name={Screens.GoldEducation}
