+++ src/transactions/feed/TransferFeedItem.tsx
@@ -17,25 +17,31 @@ import { useTokenInfo } from 'src/tokens/hooks'
 import TransactionFeedItemImage from 'src/transactions/feed/TransactionFeedItemImage'
 import { useTransferFeedDetails } from 'src/transactions/transferFeedUtils'
 import { TokenTransfer } from 'src/transactions/types'
-import { isJumpstartTransaction } from 'src/transactions/utils'
-
 interface Props {
   transfer: TokenTransfer
 }
 
 function TransferFeedItem({ transfer }: Props) {
   const { amount } = transfer
+  const isJumpstart = isJumpstartTransaction(transfer)
 
   const openTransferDetails = () => {
-    // Navigate to JumpstartTransactionDetailsScreen
-    navigate(Screens.TransactionDetailsScreen, { transaction: transfer })
+    if (isJumpstart) {
+      navigate(Screens.JumpstartTransactionDetailsScreen, { transaction: transfer })
+    } else {
+      navigate(Screens.TransactionDetailsScreen, { transaction: transfer })
+    }
+
     ValoraAnalytics.track(HomeEvents.transaction_feed_item_select)
   }
 
   const tokenInfo = useTokenInfo(amount.tokenId)
   const showTokenAmount = !amount.localAmount && !tokenInfo?.priceUsd
 
-  const { title, subtitle, recipient, customLocalAmount } = useTransferFeedDetails(transfer)
+  const { title, subtitle, recipient, customLocalAmount } = useTransferFeedDetails(
+    transfer,
+    isJumpstart
+  )
 
   const colorStyle = new BigNumber(amount.value).isPositive() ? { color: colors.primary } : {}
 
@@ -48,7 +54,7 @@ function TransferFeedItem({ transfer }: Props) {
           recipient={recipient}
           status={transfer.status}
           transactionType={transfer.__typename}
-          isJumpstart={isJumpstartTransaction(transfer)}
+          isJumpstart={isJumpstart}
         />
         <View style={styles.contentContainer}>
           <Text style={styles.title} testID={'TransferFeedItem/title'} numberOfLines={1}>
@@ -83,6 +89,13 @@ function TransferFeedItem({ transfer }: Props) {
       </View>
     </Touchable>
   )
+}
+
+function isJumpstartTransaction(tx: TokenTransfer) {
+  const jumpstartAddress = '0x7bf3fefe9881127553d23a8cd225a2c2442c438c'
+  // const jumpstartAddress = getDynamicConfigParams(DynamicConfigs[StatsigDynamicConfigs.WALLET_JUMPSTART_CONFIG])
+  //   .jumpstartContracts[tx.networkId]?.contractAddress
+  return tx.address === jumpstartAddress
 }
 
 const styles = StyleSheet.create({
